refactor: move MenuList and MenuListItem to partials instead of blocks
