--- src/components/common/Extensions/UserNavigation/partials/UserSubmenu/blocks/index.ts
@@ -1,2 +1,0 @@
-export { default as MenuList } from './MenuList/MenuList';
-export { default as MenuListItem } from './MenuListItem/MenuListItem';
