+++ src/components/common/Extensions/UserNavigation/partials/UserSubmenu/partials/MenuList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MenuList';
