+++ src/components/common/Extensions/UserNavigation/partials/UserSubmenu/partials/MenuListItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MenuListItem';
